Many small changes to get rid of warnings from the new version of IntelliJ.
I haven't commented on every file, only the non-obvious ones.  Generally, the
fixes are as follows ...
* The Groovy JDK extension method Matcher#getAt ("[]" for regex matches) returns
Object, but is effectively always List where we use it, so cast.
* The "implementation-class" property in each plugin's META-INF is used
internally by Gradle, so it's fine that our code doesn't reference it.
* Calls to ".each" on ConfigurationContainer are apparently ambiguous because of
some implicit way to use a "getByName" method, so cast argument to Closure.
* project.extensions.create returns the new thing, but project.extensions.add
returns null
* Some calls were relying on implicit type-casting of arguments, which is not
obvious and might change in future versions.
* Some return values were not supplied, or were unused.
* Some methods were unused.

* .idea\dynamic.xml
** Add some more useful dynamic properties.

* .idea\inspectionProfiles\Project_Default.xml
** Ignore "SyntaxError" outside the "Inspection" scope I created.  This is
mainly to ignore the "map-template.xml" file in unpacked Merurial folders, which
isn't actually an XML file at all.

* .idea\misc.xml
** Add info on Maven namespace used in ivy.xml files.

* .idea\scopes\Inspection.xml
** Fix the exclusions I wanted.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\PrerequisitesExtension.groovy
** Groovy's "==" on arrays may not do what we want; use the normal Java way.