While getting rid of IntelliJ warnings, I introduced an incorrect cast, which resulted in the "project.my" property always being null.
Another change I made in passing in that commit was incorrect.  I changed
PackageArtifactBuildScriptHandler so that, when creating a build script with a
packedDependencies block and adding entries corresponding to original source
dependencies, it would use the version number of that project currently being
build, not the "most recently published on a server somewhere" version.
However, I used "SourceDependencyHandler#getProject", which in fact returns the
*parent* project; I should have used "getSourceDependencyProject".