GR #4126 Source version info is not recorded for source dependencies which are not published
* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactHandler.groovy

It's possible to have a project with source dependencies which are never
published.  For example, they may be used to pull in developer documentation, or
it might just be that publishing is temporarily disabled for that module.

However, this is a problem for traceability to source.  For published modules,
you can trace through the dependency version information in the ivy.xml, then
look in the "build_info" folder to find source version information for that
dependency.  But there's nothing for unpublished dependencies.

So, this commit adds a "build_info/source_dependencies" folder which contains a
sub-folder for each source dependency in the transitive graph from that project,
giving the repo URL, repo version, and relative path at which it should be
checked out on disk.

Also, we change the actual packaging task to take all files under build_info,
not just the top-level "*.txt".

* integration\src\test\groovy\holygradle\BasicIntegrationTest_tAPIT_stdout.ok.txt
* integration\src\test\groovy\holygradle\BasicIntegrationTest_tBPC_stdout.ok.txt
** createPublishNotes now appears in the top-level list.