GR #4125 Quick attempt to improve holygradle start-up time
* intrepid-plugin\src\main\groovy\holygradle\dependencies\PackedDependencyHandler.groovy
** Removed shouldPublishDependency, because it's unused, undocumented, and bad
for traceability.  (Not publishing *source* dependencies is okay, because we
capture their VCS versions for traceability anyway.  Packed dependencies you want
to keep "secret" should just be private.)

* intrepid-plugin\src\main\groovy\holygradle\publishing\DefaultPublishPackagesExtension.groovy
** While making changes for #4126, I realised that I should prevent publishing if
a module has any packedDependencies which had noCreateSymlinkToCache called on
them, because Gradle/intrepid then can't know the version of the dependency.
** In passing, removed use of shouldPublishDependency.