GR #4071 Add test for unpacking with conflicting versions
During work on the above ticket, I introduced and then fixed a crash when there
were conflicting versions of a library in the (unresolved) dependency graph, and
the "default fail on version conflict" was turned off.

The same cause seems unlikely to reappear, but it seems like an area which could
be fragile, and it's important to have it working for diagnosing version
conflicts, so I added a test.  In doing that, I found another bug, in
DependencySettingsExtension--I forgot that the Groovy "?:" operator falls
through for false as well as for null, so my method for turning off the default
strict version checking didn't work.  That's fixed with this commit.  I also
found that you had to do "dependencySettings.defaultFailOnVersionConflict = false"
before creating any configurations, which was irritating and not as intended, so
I fixed that by moving the resolutionStrategy changes to a projectsEvaluated
block.