GR #4546 addPackedDependency is broken in buildScript handler
Simplifying this code on review, by making some breaking changes... in code
which nobody, or almost nobody, is using, and where an alternative approach in
build scripts is easy.  Those changes are as follows.

* includeSymlinks is removed.  Only one project we know of was using it, and
they found that it didn't make sense anyway, so replaced it with code to
explicitly create symlinks as needed in particular generated build scripts (to
work around MIDL compiler limitations).  The problem was that a module could
have symlinks into folders belonging to dependencies which were not present in
the set of configurations pulled in by the generated build script (typically
because they were private).
* addPinnedSourceDependency doesn't support wildcards any more.  This was only
ever used for "include all source deps from all projects", which can now be done
like this:
** gradle.projectsEvaluated { allprojects { sourceDependencies.each { addPinnedSourceDependency it } } }

* intrepid-plugin\src\main\groovy\holygradle\dependencies\DependencyHandler.groovy
** Added comments explaining the meaning of name and targetName, since I've
often seen the former used to derive the latter in code outside this class,
which is unnecessary and error-prone.

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactBuildScriptHandler.groovy
** Made pinnedSourceDependencies a set, not a list.  I noticed some source deps
could be added more than once for the same path, which makes no sense.
** Added overloads of addPinnedSourceDependency and addPackedDependency which
take the handler directly, to support the above Gradle-y replacement for
"addPinnedSourceDependency '*'".
** Removed includeSylmlinks stuff.
** Merged findSourceDependencies and findPackedDependencies, as they're doing
the same thing, now I've removed wildcard stuff.
*** In the process, removed the odd regex-matching code, which was effectively
just getting the targetName from the dependency.
** Similarly, merged addPinnedSourceDependency and collectSourceDependenciesForPinned.

* custom-gradle-core-plugin\src\main\groovy\holygradle\custom_gradle\util\Wildcard.groovy
* custom-gradle-core-plugin\src\test\groovy\holygradle\custom_gradle\util\WildcardTest.groovy
** No longer used.  Incidentally, the implementation of Wildard#match was subtly
wrong, for '?'.