GR #3080 packageArtifacts.includeBuildScript doesn't run if no real files are included
A problem was found where a packageArtifacts.<name>.includeBuildScript block?
didn't result in a @build.gradle@ being generated.  Explicitly adding a file to
the packaging block.

This turned out to be because there's a method in PackageArtifactBuildScriptHandler
which is called to find out whether any of the methods in the handler have been
called, and it wasn't checking enough state.  It only returned true if at least
one packed or pinned source dependency had been added.  I've extended it, and
added an integration test.

I also added a check in addPackedDependency, that there's at least one
configuration specified.  As I discovered when writung the test, if you don't do
that, you get nonsense output.