GR #5099 Republishing strips most build_info
I discovered on Friday that the republishing process seemed to be stripping out
most of the contents of the "build_info" directory, which obviously defeats the
purpose of having that info in the first place.

In fact it's just deleting it entirely and re-creating a mostly-empty one.

I realised this was introduced as part of the fix for #4126, ironically --
republishing hasn't been integration-tested yet, and I'm hoping to change it
soon, so I won't bother adding tests now.

This partial fix causes the task to do nothing if the republishing task is
being run.

I'll also add some special code to the one republishing meta-package I know of,
to disable the "package*" tasks so that people don't accidcentally run them
explicitly before running "republish".