GR #4940 Add build step to fill in link text from link target.
To avoid me creating lots of links with incorrect text, and because AsciiDoc
gives unhelpful text for in-document cross-refs, I wrote classes for a custom
build step which overwrites the text inside most "a" elements with text from the
link's target, be it in the same document or another in the set.

* .hgignore
** Adjust to avoid ignoring anything with "build" in the name!
** Also ignore new intermediate output folder.

* .idea\dynamic.xml
** Improved type info in website.gradle.

* .idea\gradle.xml
* .idea\modules.xml
** Add info so IDEA can find types from the buildSrc folder when they're used in
the Gradle scripts.  (This took a while to get right, partly by accident!)

* doc\website\asciidoc.conf
** Paste in the default XHTML header, with minor changes to make sure the link
to pygments.css is valid XML.

* doc\website\docinfo.html
** Make the link to jquery-ui.css be valid XML.

* doc\website\index.asciidoc
** Remove text from links, because now it's filled in automatically.

* website.gradle
** Main change -- AsciiDoc output goes to a "raw-html" folder, then the
link-fixing code runs across all HTML files (since any of them might link to
any other).
** Other changes
*** Write a special marker file to the output folder when running in "quickly"
mode, so that when later running in "non-quickly" mode, we can be smart about
whether or not we need to delete the output folder before we start.
*** Include the "docinfo.html" in the list of build inputs.
*** Add a bit more explicit info for IDEA.

* buildSrc\buildSrc.iml
** New IDEA project.

* buildSrc\src\main\groovy\holygradle\buildHelpers\BuildContext.groovy
* buildSrc\src\main\groovy\holygradle\buildHelpers\LinkTextFiller.groovy
* buildSrc\src\main\groovy\holygradle\buildHelpers\XmlDoc.groovy
* buildSrc\src\main\groovy\holygradle\buildHelpers\XmlDocumentSource.groovy
* buildSrc\src\main\groovy\holygradle\groovy\util\LICENSE.txt
* buildSrc\src\main\groovy\holygradle\groovy\util\NOTICE.txt
* buildSrc\src\main\groovy\holygradle\groovy\util\XhtmlNodePrinter.java
** New code for filling in link text.