GR #5346 Use of packageArtifacts' includeBuildScript should include default empty settings file.
If you have a meta-package which includes an automatically-generated build
script but no source dependencies, it won't auto-generate a settings.gradle
when run.  If you happen to unzip your meta-package in a folder which has an
ancestor folder with a settings.gradle in it, then trying to run the
meta-package will use that ancestor as the root, as per Gradle default
behaviour.  This is probably not what you want in the case of a meta-package.

So, this change means that, unless the user overrides it, a default settings
file will be generated for any meta-pacakge for which includeBuildScript is
called.


* custom-gradle-core-plugin\src\test\groovy\holygradle\test\RegressionFileHelper.groovy
** In passing, remove some unneeded logging from tests (which I spotted while
debugging).

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactBaseDSL.groovy
** Add property to allow users to turn off the new feature of default creation
of a settings file.

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactBuildScriptHandler.groovy
** Make this handler implement the PackageArtifactTextFileHandler interface,
already used for the settings file and misc. text files, because that simplifies
other code.

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactDescriptor.groovy
** Factored out the collection of text files (build script, settings file, and
other text files) into a separate class, so that it can control the addition of
a default settings file.

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactHandler.groovy
** Pass the project to contained objects, so that PackageArtifactBuildScriptHandler
can implement PackageArtifactTextFileHandler.
** Pass-through implementation of new createDefaultSettingsFile property.

* intrepid-plugin\src\main\groovy\holygradle\SettingsFileHelper.groovy
** Don't write out includes if there aren't any (as will be the case in a
default settings file).

* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest.groovy
* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest\buildScriptRequired\build.gradle
** Test that a settings file is created by default.
** Test that "createDefaultSettingsFile = false" causes no file to be created.
* intrepid-plugin\src\test\groovy\holygradle\packaging\PackageArtifactBuildScriptHandlerIntegrationTest_tBSR_noCreateDefaultSettingsFile.ok.txt

* intrepid-plugin\src\main\groovy\holygradle\packaging\PackageArtifactTextFileCollector.groovy
** Factored some logic out of PackageArtifactDescriptor to here, to properly
encapsulate the textFileHandlers collection.