Publishing safety change.
Change to publishing snapshot and release versions into different repos, so that
developers can have Artifactory delete permissions (for overwrite) in the case
of SNAPSHOT versions, and different permissions (deploy, but not delete) for
releases.  Otherwise, picking the wrong version number when publishing could
overwrite old versions.

We've never got it wrong yet, but should have done this ages ago.