GR #4941 Add build dependency on credential-store for publishing website.
When I added publishing to WebDAV, I used the credential-store EXE to get the
password for publishing.  I realised I should add a build dependency on the
credential-store project, and at that point I spotted that it was always
rebuilding, even if the source was unchanged.

Eventually I worked out that it was because the version info included in the
EXE includes the date and time, and if you have uncommitted changes, that will
be different every time you edit a file in your working copy.

So, now we only include the date/time when publishing.

In passing I factored out the "are we publishing?" test into a helper method on
the project.