GR #1962 "_conflict.*" mechanism seems broken/unneeded#
This mechanism, of merging incompatible configurations in the generated ivy.xml,
was always a design mistake.  If you're not publishing your result into a repo,
then you don't care what the ivy.xml looks like.  If you are, then it needs to
be usable, and it will look correct; that is, it correctly shows that you need
multiple separate configurations because dependencies have incompatible versions.