GR #5759 Don't overwrite the settings file completely each time.
This will allow users to set the project name manually in the settings file.
Once we get to Gradle 1.6, we can just do this with a plugin, but for now we'll
append some helper code, and keep the list of subprojects in a separate file.

This code will auto-upgrade (i.e., replace) settings.gradle files generated by
previous versions of the Holy Gradle.