While testing, I fixed a long-standing irritation with the "SNAPSHOT" code. If you try to run against a new SNAPSHOT version of some plugins, and you have also updated custom-gradle-core-plugin, the new version of that isn't downloaded, because it isn't marked as changing, because it's not an explicitly declared dependency of the project.
My hacky fix is to explicitly add custom-gradle-core-plugin if it's not already
added and we're in SNAPSHOT mode.

I think there may be a better way to do this in Gradle 2.x because it has more
control over dependency resolustion overrides.  Depends whether it allows you to
force an indirect dependency to be treated as changing, though.