GR #5759 Don't overwrite the settings file completely each time.
Since the settings file is no longer complete auto-generated, we should include
it in the buildScript ZIP.  This is particularly important for packaging
third-party modules -- in that case the build script may not be committed to
source control, with the version published in this ZIP file being the only copy.

* intrepid-plugin/src/main/groovy/holygradle/packaging/PackageArtifactHandler.groovy
** Include the settings file as well as the build file.

* intrepid-plugin/src/test/groovy/holygradle/packaging/PackageArtifactsIntegrationTest.groovy
** Test for the build and settings files being in the ZIP.

* publish.gradle
** Previous changes caused tagging of SNAPSHOT versions -- herewith fixed.  Also
improved the tag message to include the tag name.