GR #5759 Don't overwrite the settings file completely each time.
If a developer checks out a new copy of a project with a 'settings.gradle' which
includes the code from SettingsFileHelper, then the previous version would fail
because 'settings-subprojects.txt' didn't exist (yet).  This new version assumes
an empty set of subprojects in that case.

Making this change forced me to write the logic to upgrade code within a section
of the settings file.  I should really write a proper integration test for this
:-/

I also took the opportunity to make the generated code actually use the name of
the settings file in constructing the name of the subprojects file, which is
what happens when the two files are written.  This is mainly for testability but
also allows for alternate settings files (with alternate subproject lists).