Fixed a bug to allow resolutionStrategy.force to work.
The Holy Gradle creates separate configurations to locate the ivy.xml for each
module attached to a configuration.  However, it was creating them "from
nothing", so they would ignore any resolutionStrategy set on the original conf.
Now it copies the original and replaces the dependencies, so resolutionStrategy
configuration such as "force" will work.  (Previously it would just pick the
lexicographically newest version in some sense, so a "force" might appear to
have worked if you happened to be forcing a module to a newer version -- but
that would just have been coincidence.)