GR #6266 Auto-import site-specific HTTPS root certs.
Some organisations have Artifactory servers accessed over HTTPS, where the SSL
certificates are self-signed, or chain up to a root cert which is self-signed.
Some Windows apps can be set up to trust this certificate through Windows
mechanisms such as group policies, but Java apps have to have the certificate
explicitly added to a "trust store".  You can manually import it to the default
trust store for a given JRE installation, but you have to do it again every time
you upgrade Java, and again if you have multiple Java installations.

This task is to auto-import some certificates into a separate trust store within
the gradle wrapper folder, combined with the JRE default ones, and run Gradle
with that.

As a precursor step, this commit is moving the location for organisation-specific
files which are built into the wrapper, since there are going to be a few more
of them.