GR #6072 local_artifacts should contain build_info
Modified the CollectDependencies Task to also include the build info so there is
a record of source code versions.

Updated the build info task to expose its output directory so it can be
referenced from the collect dependencies task.

Added a test for the basic build info files to both the
PackageArtifactsIntegrationTest and CollectDependenciesIntegrationTest.