GR #5714 Replace TaskDependencies with Gradle's built-in mechanism
I discovered recently that the change I made doesn't actually work.  The "-a"
flag by default only prevents linking the (build) output of each configuration
in subprojets, not cross-project task dependencies generally.  I've added code
to explicitly check for that argument and only link the tasks if it's not there.
I also improved the task descriptions.