GR #6382 Create zipDependencies task
Creates a task zipDependencies which is similar to collectDependencies but just
creates a ZIP file directly.  In the process, I changed how local_artifacts is
detected and used, such that it's no longer necessary to overwrite the
gradle-wrapper.properties when collecting local_artifacts; instead gw.bat works
it out for you.  This means that the same source can be used with or without a
local_artifacts folder, with no changes required to the checked-in source.

* .idea\dynamic.xml
** Add defaule "ext" extension for ExtensionAware things.

* gw.bat
* https-setup-kit\gw.bat
* wrapper-starter-kit\gw.bat
** Re-arrange how the gradle-wrapper.properties is constructed to make that
process set up the distributionUrl to point into local_artifacts if it exists.
Originally that was done manually by the collectDependencies task.  Doing it
that way doesn't make sense with zipDependencies, because you don't necessarily
want to copy the wrapper into the zipped dependencies.  However, when the
original design was done, we weren't creating the gradle-wrapper.properties on
the fly, so there wasn't an obvious place to do it on start-up.  Now it's fairly
straightforward (as much as DOS batch files ever are).
** Just always setlocal and endlocal, since we're not supporting Win9x etc.
** Don't need a "\" after uses of %APP_HOME%

* integration\src\test\groovy\holygradle\BasicIntegrationTest_tAPIT_stdout.ok.txt
* integration\src\test\groovy\holygradle\BasicIntegrationTest_tBPC_stdout.ok.txt
** New task.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\CollectDependenciesHelper.groovy
** Pulled out most of the code into a class to be shared between
CollectDependenciesTask and ZipDependenciesTask.  Can't use a common base class
because they need to inherit from other classes.  (Could have used meta-class
mixin stuff, I suppose.)
** Dependency info isn't needed outside of lazyConfigration scope, so no point
pulling it out beforehand.
* No longer re-write gradle-wrapper.properties.
** Stop checking whether createPublishNotes exists -- it always will.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\CollectDependenciesTask.groovy
** Mostly factored out into CollectDependenciesHelper.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\DependenciesStateHandler.groovy
** Replace "local_artifacts" string literals with a constant.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\ZipDependenciesTask.groovy
** Simple variant of CollectDependenciesTask.

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** Add ZipDependenciesTask.
** Configure it and CollectDependenciesTask early, since they do lazy config
internally.

* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTest.groovy
* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTest\dummy-gradle-wrapper.properties
* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTest\useOnlyEmptyConfigs\build.gradle
* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTest\useOnlyViaEmptyConfigs\build.gradle
* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTestBase.groovy
* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTestBase\useOnlyEmptyConfigs\build.gradle
* intrepid-plugin\src\test\groovy\holygradle\dependencies\CollectDependenciesIntegrationTestBase\useOnlyViaEmptyConfigs\build.gradle
* intrepid-plugin\src\test\groovy\holygradle\dependencies\ZipDependenciesIntegrationTest.groovy
** Refactored and re-wrote integration test to work for either folder or ZIP.