GR #6382 Create zipDependencies task
As an added bonus, change handling of "local_artifacts" so that it works if that
folder appears in the root project folder _or_ any ancestor of that folder.  It
is now easy to share the same "offline repo server" between several projects.

* custom-gradle\src\initScripts\holy-gradle-init.gradle
** Look for local_artifacts in current folder and all ancestors.

* gw.bat
* https-setup-kit\gw.bat
* wrapper-starter-kit\gw.bat
** Look for local_artifacts in current folder and all ancestors.

* intrepid-plugin\src\main\groovy\holygradle\dependencies\CollectDependenciesHelper.groovy
** Fix bug with testing of snapshot distributions with local_artifacts.  The offline version would be
copied to a folder with the wrong name.