GR #6570 Use directory junctions instead of symlinks by default, for Win 8+
Unfortunately, jna-platform 3.4.0 is incompatible with jna-3.2.7, and the latter
is used internally by Groovy.  Adding a dependency on jna-3.2.7 gets me two
different versions of the JNA JAR on the Groovy classpath, rather than
overriding the version Groovy uses, presumably because the Groovy classpath
isn't subject to the normal rules of dependency resolution and override, because
it's built in to Gradle.  Having two versions of a JAR on the same classpath is
a bad idea, so instead I've back-ported the necessary parts from
jna-platform-3.4.0 (since it's LGPL).

This first commit is to add the original files from the jna-platform-3.4.0
sources.  The next commit will tweak them to get my code to compile.  After that
I could maybe strip out all the bits I don't need, but that might be more effort
than it's worth.