GR #6585 Add support for creating and mapping sets of related configurations.
Replace String with Configuration for mapping from a single configuration to
a configuration set (type), because it could mislead users into thinking they
can map from "a configuration set by name".

The only "trick" here is in DefaultConfigurationSetType, where
getDefaultMappingsTo has to extract the name from the Configuration to pass to
the Closure.