GR #6626 Stop fetching ivy.xml files eagerly.
Hit a problem in PackedDependenciesStateHandler. The refactoring of traversal of
packed dependencies from 6bb034d23fcc to f55df7416c24 for GR #3813, GR #4071
wasn't quite right. It looks like the rule was intended to be as follows.

1. Visit a module if
1.1. we haven't seen it already for the given configuration; AND
1.2. it's not part of the build (i.e., a source dep), because then we don't want an UnpackModuleVersion; AND
1.3. we have an ivy.xml file, because we need that for the relativePath.

2 Visit a module's children if
2.1. we haven't seen it already for the given configuration; AND
2.2. EITHER
2.2.1. it's part of the build (i.e., a source dep), because then we still want to find transitive dependencies; OR
2.2.2. (it's a packed dep and) we have an ivy.xml file, because we need that for the relativePath.

However, the "EITHER/OR" in 2.2 became an "AND", and source deps never have
ivy.xml files, so condition 2.2.1 effectively got removed, so we would only be
collecting direct and transitive packed deps of the current project, but not
those of source dependencies.

I think that's actually the right semantics, but it doesn't match the comments.
For the current definition of CollectDependenciesTask this ends up working out
okay, because we collect separately for each project (to be sure we get any
dependencies from non-linked configurations). For creating symlinks to the
cache, that's also the right thing to do, because we only want symlinks for
(direct and transitive) deps of the current project, not subprojects.

I think the things which use this need proper tests. There's a
CollectDependenciesIntegrationTest but I don't think it covers this distinction.