GR #5818 New Artifactory plugin -- fail replication if source server is different version.
In email conversation with JFrog I was told that they test replicating from
older servers to newer, because that's one possible upgrade path, but they don't
test or support replicating from newer servers to older.  Since I'm cautious,
I wrote this plugin to automatically cause all replication to fail if the
source (remote) server is a different version.  It will then email a list of
addresses as well as log an error.  Getting the email settings is fiddly, and
there's no way to get a list of all admin users to email them, so we need an
explicit list.