GR #4126 Make sure we really visit all source dependencies recursively.
The hierarchy of source dependencies is not necessarily the same as the
hierarchy of projects.  Before this change, if

* a non-root project A had a source dependency on another non-root project B, and
* the root project R only depended on A,

then

* R would list A, B, and all their source dependencies recursively, but
* A would only list its own immediate source dependencies, not those of B
recursively.

This occured because

* all source dependency projects are added as direct subprojects of the root
project, by the code injected into settings.gradle from
SettingsFileHelper.gradle -- so that source dependencies can be a DAG, not just a
(project-)tree -- and
* the code in PackageArtifactHandler which collects the source information was
only recording directly source dependencies plus those of subprojects (of which
there are none, except for the root project).

Instead, PackageArtifactHandler should (and now does) visit the projects
attached to each source dependency.

PackageArtifactBuildScriptHandler was also mis-using subprojects for source
dependncies, and had some other non-sensical logic in the same area to deal with
having multiple source dependencies with the same folder name but under
different paths, so I changed that.  The new behaviour (if I got it right) might
still be surprising and annoying, but at least it should make sense.

This still needs tests, but I wanted to push the fix so that others can use it
to make progress now.