GR #6675 Unit test plugin improvements
* idea/dynamic.xml
** Add dynamic info for "tests" collection extension.

* integration/src/test/groovy/holygradle/BasicIntegrationTest/tBPC/build.gradle
** Update to use new recommended methods and paths.

* unit-test-plugin/src/main/groovy/holygradle/unit_test/TestHandler.groovy
** Deprecate redirectOutputToFile method in favour of something closer to
ExecSpec.  (We can't just use ExecSpec because the values are templated, so we
can set a template for setStandardOutput to use across all tests, but can't add
a getStandardOutput method which would return a single OutputStream.)
** Change teeStandardOutput to setStandardOutputTee, so it's easy to switch
between that and just setStandardOutput in a build script.
** Document new setters.
** Refactor the logic for setting the task's output stream to be more readable,
with less nesting and simpler control flow.