GR #6675 Bug fix, and add support for redirecting stderr as well.
Users requested the ability to redirect standard error as well as standard
output, so I've added that.

Also users pointed out a bug -- the output stream was created at configuration
time, meaning that any run of any task would overwrite existing test logs with
an empty file!  Now, creating and setting the streams is done as a
lazyConfiguration block.  As a safety measure, the override is not done )and a
warning is logged) if something else had already reconfigured the relevant
stream to something other than its default.