GR #6588 Fixes for upgrading from Artifactory 3.3.0 to 4.7.1.
Should still work with 3.3.0 but I didn't test that.

* utils/artifactory-plugins/src/main/groovy/replicationVersionCheck.groovy
** We now need to explicitly use the proxy for a remote repo to connect to the
remote server to find out its version.
** The comparison logic was flawed -- it would complain if (A) the main version
was greater OR (B) the number of components in the main version length was OR
(C) the revision was greater.  It should complain only if
*** (A) is greater
*** OR (A) is equal and (B) is greater
*** OR (A) and (B) are equal and (C) is greater
** Artifactory 3.3.0 passes the name of the remote repo to the
beforeReplicateXxx hooks, but 4.7.1 passes the "-cache" repo name; so we now
strip off that suffix if it exists.