GR #6588 Fixes for upgrading from Artifactory 3.3.0 to 4.7.1
* intrepid-plugin/src/main/groovy/holygradle/ArtifactoryHelper.groovy
** artifactExists method was sending the wrong content type for the Accept
header.  Artifactory 3.3.0 ignored this and replied with HTML; 4.7.1 only
returns JSON and complains "406 Not Acceptable".  In fact we only need a HEAD
request and don't care about the conten, so can use default of ANY.
** Also tidied up initialisation of auth headers to avoid IntelliJ warnings.

* intrepid-plugin/src/main/groovy/holygradle/publishing/CheckPublishedDependenciesTask.groovy
** While investigating this, I found that requesting a folder (as we are doing
in this case) without the trailing '/' causes a wasted round trip because
Artifactory responds with a "302 Found" redirect whcih adds a trailing slash.