GR #4107 Fixes to get unit tests to pass (configuration set regression tests)
* intrepid-plugin/src/main/groovy/holygradle/artifacts/DefaultConfigurationSet.groovy
** The upgrade has changed the behaviour of
com.google.common.collect.Sets#cartesianProduct so that it returns items in a
different order.  Therefore, we now explicitly sort them into the order we want
-- which is to say, the order we originally had.

* intrepid-plugin/src/main/groovy/holygradle/artifacts/DefaultConfigurationSet.groovy
** If a configuration set type was created using GStrings instead of Strings,
the previous code wouldn't have a problem but, now that we're trying to compare
strings to sort them, we need to make sure they're all really Strings.  Rass'n
frass'n not-always-working dynamic type magic ...