GR #6279
* intrepid-plugin\src\main\groovy\holygradle\dependencies\SourceOverrideHandler.groovy
** Rename hasCreatedDummyModule for consistency with method names.
** Add lazy-init flag for generating the dependency file, too.
** Throw if the "from" location isn't set.

* intrepid-plugin\src\main\groovy\holygradle\IntrepidPlugin.groovy
** Fail if source overrides are added in a non-root project, because it would be
even more effort to tie together overrides across subprojects and cross-check
them.  We'd like to support that some day, maybe.  We could just not add the DSL
handler to non-root projects, but then the user would only get a
MissingMethodException instead of a useful error message.
** Don't create and add the dummy repo unless there are actually any source
overrides.  If we're adding it, add it first, because it's going to have only a
few things in it and they definitely won't be in normal repos.
** Lift the "eachDependency" call (which substitutes in the dummy source
override module versions) out of the inner loop, because it registers a lazy
callback, rather than looping over each dependency immediately.  When we had it
in the inner loop, it was adding the same callback multiple times, wasting time.
** Cache the map of all transitive dependencies for each handler, rather than
re-building it for each configuration, since we currently apply the same source
overrides in all configuations.  Same for the list of conflicts between
transitive dependencies in different source overrides.
** Optimised the check for conflicts between transitive dependencies of
different source overrides -- it no longer checks "A vs A", nor checks "B vs A"
when it has already checked "A vs B".
** Factor various steps out into methods.

* intrepid-plugin\src\main\groovy\holygradle\unpacking\UnpackModuleVersion.groovy
** getUniqueValue can't be static as it calls getFullCoordinate.

* intrepid-plugin\src\test\groovy\holygradle\dependencies\ReplaceWithSourceIntegrationTest.groovy
** Added explanatory comments to some tests.
** Updated the message for incompatibleTransitiveDependencies because the code
now outputs a more detailed message; still the same meaning.
** Updated the message for incompatibleTransitiveDependencies.  The previous
meaning was different and NOT what this test should have been testing; and I'm
pretty confident that the new tested output is what's expected.  However, I'm
not clear what I changed that caused the output to be different :-/

* intrepid-plugin\src\test\groovy\holygradle\source_dependencies\ProjectDependenciesIntegrationTest.groovy
** Refactoring in IntrepidPlugin changed the output slightly.

* intrepid-plugin\src\test\groovy\holygradle\dependencies\ReplaceWithSourceIntegrationTest\ProjectKIn\build.gradle
** New test script for the case of a direct dependency which clashes with an
indirect one in a source replacement.